Split NavButtons into a separate component.
